Convert whitelist integration test to RSpec
