+++ features/step_definitions/configure_whitelist_steps.rb
@@ -1,11 +1,11 @@
 Given(/^I have an app that depends on an MIT license$/) do
-  @user = ::DSL::User.new
+  @user = LicenseFinder::TestingDSL::User.new
   @user.create_ruby_app
   @user.create_and_depend_on_gem 'mit_gem', license: 'MIT'
 end
 
 Given(/^I have an app that depends on an BSD license$/) do
-  @user = ::DSL::User.new
+  @user = LicenseFinder::TestingDSL::User.new
   @user.create_ruby_app
   @user.create_and_depend_on_gem 'bsd_gem', license: 'BSD'
 end
