+++ features/step_definitions/node_steps.rb
@@ -1,5 +1,5 @@
 Given(/^A package file with dependencies$/) do
-  @user = ::DSL::User.new
+  @user = LicenseFinder::TestingDSL::User.new
   @user.create_node_app
 end
